requirements.txt
Django>=4.2
pillow
psycopg2-binary
